requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
requests 
python-multipart
pydantic